dashboard-agent/pyproject.toml
[project]
name = "dashboard-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "Pillow>=10.0",
]

[project.scripts]
dashboard-agent = "src.orchestrator:main"
